+++ api/routes/admin.js
@@ -6,11 +6,7 @@ const clientHandler = require('../handlers/admin_client');
 
 
 router.get("/order/get", handler.getOrders);
-router.get("/order/get/:id",handler.getOrder);
-router.post("/order/create", handler.addOrder);
-router.delete("/order/delete", handler.deleteOrder);
-// router.delete("/orderTemp/delete", handler.deleteOrderTemp);
-// router.get("/orderTemp/get", handler.getOrderTemp);
+router.get("/order/get/:value",handler.getOrder);
 
 router.get("/verifiedClients", clientHandler.getVerifiedClients);
 router.get("/notVerifiedClients", clientHandler.getNotVerifiedClients);
