+++ api/routes/index.js
@@ -8,6 +8,7 @@ const adminRoutes = require("./admin")
 const emailRoutes = require("./email")
 const searchRoutes = require("./search")
 const imageRoutes = require("./image_upload")
+const clientRoutes = require("./client")
 
 router.get('/', function (req, res){
     console.log(__dirname)
@@ -21,5 +22,6 @@ router.use('/admin',adminRoutes);
 router.use('/email',emailRoutes);
 router.use('/search/',searchRoutes);
 router.use('/image', imageRoutes);
+router.use('/client', clientRoutes);
 
 module.exports = router;
